Upgrade Android dependencies and improve UI
This commit upgrades various Android dependencies, including Kotlin,
Android Gradle Plugin, and WebRTC library.

It also introduces UI improvements:
- Adds a Toolbar to the main activity.
- Uses Material Components theme.
- Updates the app icon to an adaptive icon.
- Improves the layout of the chat interface.
- Changes the color of debug messages in the console.
- Adds build features configuration to `gradle.properties`.
- Removes unnecessary blank lines in Kotlin files.

+++ app/src/main/kotlin/cz/sazel/android/serverlesswebrtcandroid/MainActivity.kt
@@ -1,22 +1,34 @@
 package cz.sazel.android.serverlesswebrtcandroid
 
+import android.graphics.Color
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
-import android.view.View.GONE
+import android.view.View.INVISIBLE
 import android.view.View.VISIBLE
 import android.widget.Toast
+import androidx.activity.SystemBarStyle
+import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.ActivityCompat
+import androidx.core.graphics.Insets
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import cz.sazel.android.serverlesswebrtcandroid.console.RecyclerViewConsole
 import cz.sazel.android.serverlesswebrtcandroid.databinding.ActivityMainBinding
 import cz.sazel.android.serverlesswebrtcandroid.webrtc.ServerlessRTCClient
-import cz.sazel.android.serverlesswebrtcandroid.webrtc.ServerlessRTCClient.State.*
-
+import cz.sazel.android.serverlesswebrtcandroid.webrtc.ServerlessRTCClient.State.CHAT_ENDED
+import cz.sazel.android.serverlesswebrtcandroid.webrtc.ServerlessRTCClient.State.CHAT_ESTABLISHED
+import cz.sazel.android.serverlesswebrtcandroid.webrtc.ServerlessRTCClient.State.CREATING_ANSWER
+import cz.sazel.android.serverlesswebrtcandroid.webrtc.ServerlessRTCClient.State.CREATING_OFFER
+import cz.sazel.android.serverlesswebrtcandroid.webrtc.ServerlessRTCClient.State.INITIALIZING
+import cz.sazel.android.serverlesswebrtcandroid.webrtc.ServerlessRTCClient.State.WAITING_FOR_ANSWER
+import cz.sazel.android.serverlesswebrtcandroid.webrtc.ServerlessRTCClient.State.WAITING_FOR_OFFER
+import cz.sazel.android.serverlesswebrtcandroid.webrtc.ServerlessRTCClient.State.WAITING_TO_CONNECT
 
 class MainActivity : AppCompatActivity(), ServerlessRTCClient.IStateChangeListener, ActivityCompat.OnRequestPermissionsResultCallback {
-
 
     lateinit var console: RecyclerViewConsole
 
@@ -26,21 +38,26 @@ class MainActivity : AppCompatActivity(), ServerlessRTCClient.IStateChangeListen
 
     private var retainInstance: Boolean = false
 
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        enableEdgeToEdge(statusBarStyle = SystemBarStyle.light(scrim = Color.argb(0xe6, 0xFF, 0xFF, 0xFF), darkScrim = Color.argb(0x80, 0x1b, 0x1b, 0x1b)))
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
+        setSupportActionBar(binding.toolbar)
+        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
+            val typeMask: Int = WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.displayCutout() or WindowInsetsCompat.Type.ime()
+            val bars: Insets = insets.getInsets(typeMask)
+            v.setPadding(bars.left, bars.top, bars.right, bars.bottom)
+            WindowInsetsCompat.CONSUMED
+        }
         val layoutManager = LinearLayoutManager(this)
-
         binding.recyclerView.layoutManager = layoutManager
         layoutManager.stackFromEnd = true
-
+        val dividerItemDecoration = DividerItemDecoration(this, layoutManager.orientation)
+        binding.recyclerView.addItemDecoration(dividerItemDecoration)
         binding.apply {
             console = RecyclerViewConsole(recyclerView)
             console.initialize(savedInstanceState)
-
-
             val retainedClient = lastCustomNonConfigurationInstance as ServerlessRTCClient?
             if (retainedClient == null) {
                 client = ServerlessRTCClient(console, applicationContext, this@MainActivity)
@@ -54,7 +71,6 @@ class MainActivity : AppCompatActivity(), ServerlessRTCClient.IStateChangeListen
                 client = retainedClient
                 onStateChanged(client.state)
             }
-
             btSubmit.setOnClickListener { sendMessage() }
             edEnterArea.setOnEditorActionListener { _, _, _ ->
                 sendMessage()
@@ -62,7 +78,6 @@ class MainActivity : AppCompatActivity(), ServerlessRTCClient.IStateChangeListen
             }
         }
     }
-
 
     private fun sendMessage() {
         binding.apply {
@@ -87,18 +102,16 @@ class MainActivity : AppCompatActivity(), ServerlessRTCClient.IStateChangeListen
         return client
     }
 
-
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
         menuInflater.inflate(R.menu.menu, menu)
         mnuCreateOffer = menu?.findItem(R.id.mnuCreateOffer)
         return true
     }
 
-    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
-        when (item?.itemId) {
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
             R.id.mnuCreateOffer -> client.makeOffer()
         }
-
         return super.onOptionsItemSelected(item)
     }
 
@@ -107,13 +120,12 @@ class MainActivity : AppCompatActivity(), ServerlessRTCClient.IStateChangeListen
         console.onSaveInstanceState(outState)
     }
 
-
     override fun onStateChanged(state: ServerlessRTCClient.State) {
         //it could be in different thread
         binding.apply {
             runOnUiThread {
                 edEnterArea.isEnabled = true
-                progressBar.visibility = GONE
+                progressBar.visibility = INVISIBLE
                 mnuCreateOffer?.isVisible = false
                 when (state) {
                     CHAT_ENDED, INITIALIZING -> client.waitForOffer()
@@ -134,9 +146,7 @@ class MainActivity : AppCompatActivity(), ServerlessRTCClient.IStateChangeListen
     }
 
     override fun onDestroy() {
-        if (!retainInstance)
-            client.destroy()
+        if (!retainInstance) client.destroy()
         super.onDestroy()
-
     }
 }
